@@ -1,5 +1,5 @@
 {
-    "name": "udacity-c3-frontend",
+    "name": "udagram-frontend",
     "version": "0.0.1",
     "author": "Ionic Framework",
     "homepage": "https://ionicframework.com/",
@@ -24,12 +24,14 @@
         "@ionic-native/splash-screen": "^5.0.0",
         "@ionic-native/status-bar": "^5.0.0",
         "@ionic/angular": "^4.1.0",
+        "@ionic/cli": "^6.16.3",
         "core-js": "^2.5.4",
+        "fsevents": "^2.3.2",
         "rxjs": "~6.5.4",
         "zone.js": "~0.9.1"
     },
     "devDependencies": {
-        "@angular-devkit/architect": "^0.1200.5",
+        "@angular-devkit/architect": "~0.12.3",
         "@angular-devkit/build-angular": "^0.803.24",
         "@angular-devkit/core": "~7.2.3",
         "@angular-devkit/schematics": "~7.2.3",
@@ -54,7 +56,7 @@
         "protractor": "~5.4.0",
         "ts-node": "~8.0.0",
         "tslint": "~5.12.0",
-        "typescript": "~3.4.5"
+        "typescript": "^3.5.3"
     },
     "description": "An Ionic project"
 }
